+++ cs/MarkdownProcessor/Tags/HighlightTextPairTag.cs
@@ -1,4 +1,4 @@
-﻿namespace MarkdownProcessor.Markdown;
+﻿namespace MarkdownProcessor.Tags;
 
 public abstract class HighlightTextPairTag : ITag
 {
