+++ cs/MarkdownProcessor/Tags/ITag.cs
@@ -1,4 +1,4 @@
-﻿namespace MarkdownProcessor.Markdown;
+﻿namespace MarkdownProcessor.Tags;
 
 public interface ITag
 {
