+++ cs/MarkdownProcessor/Tags/Italic.cs
@@ -1,4 +1,6 @@
-﻿namespace MarkdownProcessor.Markdown;
+﻿using MarkdownProcessor.Tags;
+
+namespace MarkdownProcessor.Tags;
 
 public class Italic : HighlightTextPairTag, ITag
 {
